+++ changelog.md
@@ -10,4 +10,4 @@
 * change 'translations' property to 'interwiki'
 * support {{main}} and {{wide image|}} templates
 ## 2.1.0
-* support table '! row' row heading syntax
+* support table '! row' row heading syntax, and other forms